Create code 1
This code defines a function called getGMT0Time that creates a new Date object for the current time, and then creates a new Date object for the GMT+0 time by adding the timezone offset (in minutes) to the current time (converted to milliseconds). The console.log statement at the end of the function call the function and logs the GMT+0 time to the console.